+++ src/GRA.Controllers/ProfileController.cs
@@ -361,6 +361,21 @@ namespace GRA.Controllers
                 viewModel.ActivityDescriptionPlural = pointTranslation.ActivityDescriptionPlural;
             }
 
+            var askPersonalPointGoal = await GetSiteSettingBoolAsync(
+                SiteSettingKey.Users.AskPersonalPointGoal);
+            if (askPersonalPointGoal)
+            {
+                viewModel.AskPersonalPointGoal = true;
+                viewModel.MinimumPersonalPointGoal = programList.First().AchieverPointAmount;
+
+                var (maximumPointsSet, maximumPointsValue) = await GetSiteSettingIntAsync(
+                            SiteSettingKey.Users.MaximumActivityPermitted);
+                if (maximumPointsSet)
+                {
+                    viewModel.MaximumPersonalPointGoal = maximumPointsValue;
+                }
+            }
+
             return View("HouseholdAdd", viewModel);
         }
 
@@ -411,11 +426,15 @@ namespace GRA.Controllers
             var (askActivityGoal, defaultDailyGoal) = await GetSiteSettingIntAsync(
                 SiteSettingKey.Users.DefaultDailyPersonalGoal);
 
+            var askPersonalPointGoal = await GetSiteSettingBoolAsync(
+                SiteSettingKey.Users.AskPersonalPointGoal);
+
             bool askAge = false;
             bool askSchool = false;
+            Program program = null;
             if (model.User.ProgramId >= 0)
             {
-                var program = await _siteService.GetProgramByIdAsync(model.User.ProgramId);
+                program = await _siteService.GetProgramByIdAsync(model.User.ProgramId);
                 askAge = program.AskAge;
                 askSchool = program.AskSchool;
                 if (program.AgeRequired && !model.User.Age.HasValue)
@@ -493,6 +512,21 @@ namespace GRA.Controllers
                         model.User.DailyPersonalGoal = null;
                     }
 
+                    if (!askPersonalPointGoal
+                        || model.User.PersonalPointGoal <= program.AchieverPointAmount)
+                    {
+                        model.User.PersonalPointGoal = null;
+                    }
+                    else
+                    {
+                        var (maximumPointsSet, maximumPointsValue) = await GetSiteSettingIntAsync(
+                            SiteSettingKey.Users.MaximumActivityPermitted);
+                        if (maximumPointsSet && model.User.PersonalPointGoal > maximumPointsValue)
+                        {
+                            model.User.PersonalPointGoal = maximumPointsValue;
+                        }
+                    }
+
                     var newMember = await _userService.AddHouseholdMemberAsync(authUser.Id,
                         model.User,
                         false,
@@ -552,6 +586,21 @@ namespace GRA.Controllers
                     .Replace("{0}", "", StringComparison.OrdinalIgnoreCase)
                     .Trim();
                 model.ActivityDescriptionPlural = pointTranslation.ActivityDescriptionPlural;
+            }
+
+            if (askPersonalPointGoal)
+            {
+                model.AskPersonalPointGoal = true;
+
+                var (maximumPointsSet, maximumPointsValue) = await GetSiteSettingIntAsync(
+                    SiteSettingKey.Users.MaximumActivityPermitted);
+                if (maximumPointsSet)
+                {
+                    model.MaximumPersonalPointGoal = maximumPointsValue;
+                }
+
+                program ??= programList.First();
+                model.MinimumPersonalPointGoal = program.AchieverPointAmount;
             }
 
             return View("HouseholdAdd", model);
@@ -1426,6 +1475,21 @@ namespace GRA.Controllers
                 viewModel.ActivityDescriptionPlural = pointTranslation.ActivityDescriptionPlural;
             }
 
+            var askPersonalPointGoal = await GetSiteSettingBoolAsync(
+                SiteSettingKey.Users.AskPersonalPointGoal);
+            if (askPersonalPointGoal)
+            {
+                viewModel.AskPersonalPointGoal = true;
+                viewModel.MinimumPersonalPointGoal = userProgram.AchieverPointAmount;
+
+                var (maximumPointsSet, maximumPointsValue) = await GetSiteSettingIntAsync(
+                    SiteSettingKey.Users.MaximumActivityPermitted);
+                if (maximumPointsSet)
+                {
+                    viewModel.MaximumPersonalPointGoal = maximumPointsValue;
+                }
+            }
+
             return View(viewModel);
         }
 
@@ -1470,6 +1534,9 @@ namespace GRA.Controllers
             var (askActivityGoal, defaultDailyGoal) = await GetSiteSettingIntAsync(
                 SiteSettingKey.Users.DefaultDailyPersonalGoal);
 
+            var askPersonalPointGoal = await GetSiteSettingBoolAsync(
+                SiteSettingKey.Users.AskPersonalPointGoal);
+
             if (ModelState.IsValid)
             {
                 try
@@ -1512,6 +1579,21 @@ namespace GRA.Controllers
                     else
                     {
                         model.User.DailyPersonalGoal = null;
+                    }
+
+                    if (!askPersonalPointGoal
+                        || model.User.PersonalPointGoal <= program.AchieverPointAmount)
+                    {
+                        model.User.PersonalPointGoal = null;
+                    }
+                    else
+                    {
+                        var (maximumPointsSet, maximumPointsValue) = await GetSiteSettingIntAsync(
+                            SiteSettingKey.Users.MaximumActivityPermitted);
+                        if (maximumPointsSet && model.User.PersonalPointGoal > maximumPointsValue)
+                        {
+                            model.User.PersonalPointGoal = maximumPointsValue;
+                        }
                     }
 
                     await _userService.Update(model.User);
@@ -1580,6 +1662,19 @@ namespace GRA.Controllers
                     .Replace("{0}", "", StringComparison.OrdinalIgnoreCase)
                     .Trim();
                 model.ActivityDescriptionPlural = pointTranslation.ActivityDescriptionPlural;
+            }
+
+            if (askPersonalPointGoal)
+            {
+                model.AskPersonalPointGoal = true;
+                model.MinimumPersonalPointGoal = program.AchieverPointAmount;
+
+                var (maximumPointsSet, maximumPointsValue) = await GetSiteSettingIntAsync(
+                    SiteSettingKey.Users.MaximumActivityPermitted);
+                if (maximumPointsSet)
+                {
+                    model.MaximumPersonalPointGoal = maximumPointsValue;
+                }
             }
 
             return View(model);