+++ src/GRA.Controllers/ViewModel/Profile/ProfileDetailViewModel.cs
@@ -8,11 +8,14 @@ namespace GRA.Controllers.ViewModel.Profile
         public string ActivityDescriptionPlural { get; set; }
         public bool AskEmailSubscription { get; set; }
         public string AskEmailSubscriptionText { get; set; }
+        public bool AskPersonalPointGoal { get; set; }
         public SelectList BranchList { get; set; }
         public string BranchName { get; set; }
         public EmailAwardViewModel EmailAwardModel { get; set; }
         public bool HasAccount { get; set; }
         public int HouseholdCount { get; set; }
+        public int? MaximumPersonalPointGoal { get; set; }
+        public int? MinimumPersonalPointGoal { get; set; }
         public string ProgramJson { get; set; }
         public SelectList ProgramList { get; set; }
         public string ProgramName { get; set; }